tax example: fixed the calculation bug from the talk
When I showed the tax example, I had a logic error in the split_by_bracket
implementation that would record the amount for brackets in the middle
incorrectly if earnings spanned more than two brackets.  The gist  Of the
error was that I was tracking the amount remaining to be taxed when I
should have been tracking the amount that had already been taxed.  Here is
the corrected implementation.